OTA updates 5.1 LMY47O.H5 and LMY47O.H6
This represents a major change from using the community-supplied "RUU"
images to the Google-supplied "OTA" images.  The difference is that the
OTA image is incremental and must be applied on top of the version before
it.  The benefit is that I no longer have to rely on the community to
put out RUU images.